test(frontend): update snaps
