+++ packages/react-ui/src/Grid/Grid.js
@@ -1,0 +1,52 @@
+import React from "react";
+import PropTypes from "prop-types";
+import styled from "styled-components";
+import { OverflowWrapper } from "../OverflowWrapper";
+import { breakpoints, spacings } from "../theme";
+
+export const GridContext = React.createContext({ columns: 4 });
+
+export const Grid = ({ columns, ...props }) => {
+  return (
+    <StyledOverflowWrapper mobileOnly>
+      <GridContext.Provider value={columns}>
+        <List {...props} />
+      </GridContext.Provider>
+    </StyledOverflowWrapper>
+  );
+};
+Grid.propTypes = {
+  columns: PropTypes.number
+};
+Grid.defaultProps = {
+  columns: 4
+};
+
+export const StyledOverflowWrapper = styled(OverflowWrapper)`
+  @media (max-width: ${breakpoints.mobile}) {
+    margin-bottom: ${spacings.base};
+  }
+`;
+
+export const List = styled.ul`
+  display: flex; /* Flex layout so items have equal height. */
+  flex-wrap: wrap;
+  align-content: stretch;
+  align-items: stretch;
+  justify-content: flex-start;
+  /* Use negative margins on sides to create gutters that do not also
+    create a gutter at the edges of the container. */
+  margin-top: ${spacings.medium};
+  margin-right: calc(-1 * ${spacings.small});
+  margin-bottom: ${spacings.larger};
+  margin-left: calc(-1 * ${spacings.small});
+  padding: 0;
+  list-style-type: none;
+  @media (max-width: ${breakpoints.mobile}) {
+    flex-wrap: nowrap;
+    margin-top: ${spacings.small};
+    margin-right: 0;
+    margin-bottom: ${spacings.medium};
+    margin-left: 0;
+  }
+`;
