+++ packages/react-ui/src/Grid/test.js
@@ -6,14 +6,14 @@ describe("<Grid />", () => {
   test("should render", () => {
     const { container } = render(
       <Grid>
-        <GridCell>
+        <div>
           <strong>titre 1</strong>
           <p>contenu 1</p>
-        </GridCell>
-        <GridCell>
+        </div>
+        <div>
           <strong>titre 2</strong>
           <p>contenu 2</p>
-        </GridCell>
+        </div>
       </Grid>
     );
     expect(container).toMatchSnapshot();
@@ -21,14 +21,14 @@ describe("<Grid />", () => {
   test("should render 5columns", () => {
     const { container } = render(
       <Grid columns={5}>
-        <GridCell>
+        <div>
           <strong>titre 1</strong>
           <p>contenu 1</p>
-        </GridCell>
-        <GridCell>
+        </div>
+        <div>
           <strong>titre 2</strong>
           <p>contenu 2</p>
-        </GridCell>
+        </div>
       </Grid>
     );
     expect(container).toMatchSnapshot();
