+++ packages/react-ui/src/index.js
@@ -15,7 +15,7 @@ export { Alert } from "./Alert";
 export { Button } from "./Button";
 export { CardList } from "./CardList";
 export { GlobalStyles } from "./GlobalStyles";
-export { Grid, GridCell } from "./Grid";
+export { Grid, RootGrid, GridCell } from "./Grid";
 export { FlatList } from "./FlatList";
 export { Modal, ModalContentWrapper } from "./Modal";
 export { OverflowWrapper } from "./OverflowWrapper";
